requirements.txt
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
Django==1.11
django-extensions==1.9.9
django-filter==1.0.2
django-rest-swagger==2.1.2
djangorestframework==3.5.1
djangorestframework-expander==0.2.3
djangorestframework-filters==0.10.2
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
parso==0.1.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
pyreadline==2.1
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
tornado==4.5.3
traitlets==4.3.2
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
yapf==0.20.1
